requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
djangorestframework-simplejwt[crypto]>=5.3.0
django-filter>=23.0
django-cors-headers>=4.3.0
Pillow>=10.0.0
python-dateutil>=2.8.2
psycopg2-binary>=2.9.9
celery>=5.3.0
redis>=5.0.0
django-environ>=0.11.0
drf-spectacular>=0.27.0
gunicorn>=21.2.0
whitenoise>=6.6.0
